refactor: break headers into separate file
